bot/requirements.txt
trafilatura
readability-lxml
beautifulsoup4
requests
pyyaml
python-slugify
python-frontmatter
markdownify
feedparser
